Repo signature, stashed commits
Create a repo signature (a hash of the ref hashes and ref names),
to use as a proxy for "did this repo change".

Show stashed commits (stashed commits aren't normally pushed
upstream, so they are like having uncommitted or unpushed changes).

Also show elapsed time, mostly to look for things we want to make
optional.