Fix several errors with remotes
The ls-remote command could return a fingerprint mismatch error;
show this so that the local fingerprint can be updated or the
remote repo rejected as being hijacked.

Also fix error with matching refs - it's ok for a remote ref to
exist that has no local ref. Right now this is vocal, but probably
should be just normal status (maybe behind verbose output).